version 1.1
bugs opgelost:
 -Melding dat de mail wat langer kan duren.
 -Je kan nu ranges bepalen bij het aanmaken van een activiteit. Bijvoorbeeld 'Garasmaaien tussen de 10 en 30 graden'
 -De knoppen zijn vertaald naar het Nederlands
 -Titel bij het activiteiten overzicht
 -Je kan geen taken meer verwijderen via de url
 -

+++ include/delete.php
@@ -5,6 +5,21 @@ if (!isset($_SESSION['loggedin'])) {
     exit;
 }
 
+
+
+//connecting to the database
+include 'db.php';
+
+$dbhost = 'localhost';
+$dbuser = 'deb85590_buitenplanner';
+$dbpass = '$KT^8L4qiRDL!e';
+$dbname = 'deb85590_buitenplanner';
+
+$db = new db($dbhost, $dbuser, $dbpass, $dbname);
+//get activity and check
+
+
+
 if (isset($_GET['id'])) {
     $databaseHost = 'localhost';
     $databaseName = 'deb85590_buitenplanner';
@@ -13,11 +28,23 @@ if (isset($_GET['id'])) {
 
     $mysqli = mysqli_connect($databaseHost, $databaseUsername, $databasePassword, $databaseName);
 
-    // sql to delete a record
-    $sql = "DELETE FROM activity WHERE activity_ID=" . (int)$_GET['id'];
-    $result = mysqli_query($mysqli, $sql);
+    $id =  $_GET['id'];
+    $email = $_SESSION['name'];
 
+    $check = $db->query('SELECT * FROM activity WHERE activity_ID = ? AND email = ?', $id, $_SESSION['name']);
+//   echo  $check->numRows();
+
+if ($check->numRows() == 1){
+    // sql to delete a record
+    $sql = "DELETE FROM activity WHERE  activity_ID=" . (int)$_GET['id'] ;
+    $result = mysqli_query($mysqli, $sql);
+        header('Location: createActivity.php');
+}else{
+    echo 'jij mag dit niet aanpassen';
     header('Location: createActivity.php');
+
+}
+
 }
 
 else{
